Move ADJUST-URL-FOR-RAW to core library
While working around CORS problems and using the GitLab API to fetch
data for a raw URL is specific to the web REPL, another feature that was
developed for the REPL is actually more generally useful.  That is being
able to get a decorated GitHub or GitLab link and DO or IMPORT it by
navigating to the appropriate raw link for that code.

This is specific to things like DO, and not lower level features like
READ...because READ may actually want to get the HTML decoration
data and line numbers/etc. for processing.  There needs to be a way to
do that.  So it isn't a generic redirection--only for code that wants to
be a little more convenient.

The service is used by CSS-DO and JS-DO as well.