Make /AUTOMIME automatic on GitHub/GitLab
CORS rules prevent the use of css or js in script or stylesheet
conventional loading if they don't have the right MIME type.  And
GitHub and GitLab raw links don't have any MIME types.

The /AUTOMIME feature was added to just use a normal fetch()
request to get the text, and then fabricate the css and js from
that instead.  But since you always have to do that for GitHub
and GitLab links it seems better to make it automatic there.