Make stdout content always in a span
Originally, content for a line was put directly into a DIV with class
'line' as the content for that node.  This created some trouble when
input SPAN came along, because it limited the flexibility in how that
span could be styled relative to its peer content inside the line div.
This meant introducing an "input-container" div and retroactively
editing the replpad's layout.

A better strategy is likely to always have a separate styled span for
the splices of `stdout` content, so this problem does not occur.  As
it can still be merged as output arrives, this really only adds 1
extra node per line.