Tentative workaround for Android key event problems
Android using the default suggesting keyboard does not give us
keydown/keypress/keyup events...at least not most of the time.  This is
due to not wanting the user to type `btaman` and send keys like `b`
and `t` and `a` when they may take a suggestion of `batman` which
would be completely different.

So even though it puts characters on the screen in the contenteditable
it is willing to do "takebacks" on that.

Even worse it seems to do random things with sending `\n` sometimes
as part of a key sequence, but Enter at other times.

The real suggestion is to have users switch to a keyboard that mimics
a physical keyboard, and allowing us to implement our own in-REPL
complete suggestions (the same ones we'd offer web users on PCs, if
we had the feature).  The Hacker's Keyboard seems to do this.

But if at all possible we don't want to make it impossible to use the
ReplPad with the default keyboard.  This uses some heuristics based
on 'input' events separate from keydown/keypress/keyup.  The goal isn't
to be perfect, but "good enough" that people can try it and then
eventually switch over to the Hacker's Keyboard if they like what
they are seeing.

(We've discussed bundling the Hacker's Keyboard actually in with the
APK wrapper for ReplPad, which seems like an even better idea now than
it already did.)