Remove .click() call in ActivateInput
The click call required the input to be set...which drew attention to
the fact that clicking was triggering a call to the ReplPad's own
callback code.

Originally click() was done to try and get the insertion caret to show
up on Android, but it turned out to not really be related.

If it doesn't appear to be doing anything useful, it's probably best
to remove it.  But leaving behind notes in case there's some reason
for self-clicking that arises in the future.