Fix LOOP instances
For a time, WHILE was slated to become arity-1 in order to match the
meaning in PARSE.  It never happened...but for a time all uses of
WHILE had to be converted to LOOP.

This had the positive effect of calling out which WHILE were in PARSE,
because those are now deprecated to become MAYBE SOME or OPT SOME (or
if the optionality was not required, just SOME).  As those instances
are changed, LOOP can be converted back to WHILE again.

LOOP is being reserved for a generalized looping dialect, and REPEAT
is used for simple counts.  Fixes the README.md on this point.