Add EPARSE Parsing Demo Integrating w/CodeMirror
This shows how UPARSE is expanded using a custom MARK combinator
to talk to CodeMirror through an extension.  The CodeMirror
extension is loaded on demand, and the EPARSE functionality should
also be in its own .reb file (like the watchlist, but the method
used in that needs improvement in light of new module rules).