Return nothing instead of trash
Trash is now the reified quasiblank, and Nothing is the state of not
being defined.

In Ren-C you can just write `return ~` or leave the return off the
bottom of the function.  This takes advantage of JavaScript giving
undefined on RETURN with no argument and when there's no return
statement at the bottom of the function, which we interpret as
meaning NOTHING.