Don't use # for ignored multi-returns
It's now _ to opt out of naming.  The reasoning is that you no longer
"opt out" of values in general, the assignment will have the value of
the right hand side, usually a PACK!.

So the only distinction could be {_} vs. {#} ... and it's not clear
that the extra explanation and implementation is worth it to be able
to pipe a null (void? trash) out vs. the value in the spot.