Switch from .reb back to historical .r
Rebol has about the same legitimacy of "claim" to .r as the R language
(coming from using the extension widely in the same era).  It's no
longer as much of an issue in the time of AI to worry about the
extension conflict confusing syntax highlighting...so using the
less noisy name is preferable.